+++ app/src/main/java/com/jin/jjinweather/feature/temperature/ui/weathercontent/success/WeatherSuccessScreen.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.jin.jjinweather.feature.weather.domain.model.CityWeather
 import com.jin.jjinweather.ui.theme.SuccessBackgroundBottomDayColor
@@ -17,6 +18,23 @@ import com.jin.jjinweather.ui.theme.SuccessBackgroundTopDayColor
 import com.jin.jjinweather.ui.theme.SuccessBackgroundTopNightColor
 import com.jin.jjinweather.ui.theme.SuccessCardBackgroundDayColor
 import com.jin.jjinweather.ui.theme.SuccessCardBackgroundNightColor
+import com.jin.jjinweather.ui.theme.Temperature10
+import com.jin.jjinweather.ui.theme.Temperature11
+import com.jin.jjinweather.ui.theme.Temperature12
+import com.jin.jjinweather.ui.theme.Temperature13
+import com.jin.jjinweather.ui.theme.Temperature14
+import com.jin.jjinweather.ui.theme.Temperature15
+import com.jin.jjinweather.ui.theme.Temperature16
+import com.jin.jjinweather.ui.theme.Temperature17
+import com.jin.jjinweather.ui.theme.Temperature18
+import com.jin.jjinweather.ui.theme.Temperature19
+import com.jin.jjinweather.ui.theme.Temperature20
+import com.jin.jjinweather.ui.theme.Temperature21
+import com.jin.jjinweather.ui.theme.Temperature22
+import com.jin.jjinweather.ui.theme.Temperature23
+import com.jin.jjinweather.ui.theme.Temperature24
+import com.jin.jjinweather.ui.theme.Temperature25
+import com.jin.jjinweather.ui.theme.Temperature9
 import java.time.LocalTime
 
 @Composable
@@ -88,3 +106,26 @@ private fun generateBackgroundColor(isAfterSunset: Boolean) = Brush.verticalGrad
 
 private fun generatedCardBackgroundColor(isAfterSunset: Boolean) =
     if (isAfterSunset) SuccessCardBackgroundNightColor else SuccessCardBackgroundDayColor
+
+fun mapTemperatureToColor(temperature: Int): Color {
+    return when (temperature) {
+        9 -> Temperature9
+        10 -> Temperature10
+        11 -> Temperature11
+        12 -> Temperature12
+        13 -> Temperature13
+        14 -> Temperature14
+        15 -> Temperature15
+        16 -> Temperature16
+        17 -> Temperature17
+        18 -> Temperature18
+        19 -> Temperature19
+        20 -> Temperature20
+        21 -> Temperature21
+        22 -> Temperature22
+        23 -> Temperature23
+        24 -> Temperature24
+        25 -> Temperature25
+        else -> Temperature10
+    }
+}
